+++ CHANGES.rst
@@ -2,6 +2,9 @@
 CHANGES
 =======
 
+develop
+=======
+
 2015-06-25 0.1.8
 ================
 - Use `product_inv_id` for published product lookup in core products sync